etape 4 de la phase 1 avec affichage de name et package String table
