--- Examples_fusion/.deps/eabi-none-lib.Po
@@ -1,1 +1,0 @@
-# dummy
