--- Examples_fusion/config.h
@@ -1,26 +1,0 @@
-/* config.h.  Generated from config.h.in by configure.  */
-/* config.h.in.  Generated from configure.ac by autoheader.  */
-
-/* Name of package */
-#define PACKAGE "arm-examples"
-
-/* Define to the address where bug reports for this package should be sent. */
-#define PACKAGE_BUGREPORT "Guillaume.Huard@imag.fr"
-
-/* Define to the full name of this package. */
-#define PACKAGE_NAME "Arm Examples"
-
-/* Define to the full name and version of this package. */
-#define PACKAGE_STRING "Arm Examples 1.0"
-
-/* Define to the one symbol short name of this package. */
-#define PACKAGE_TARNAME "arm-examples"
-
-/* Define to the home page for this package. */
-#define PACKAGE_URL ""
-
-/* Define to the version of this package. */
-#define PACKAGE_VERSION "1.0"
-
-/* Version number of package */
-#define VERSION "1.0"
