--- Examples_fusion/stamp-h1
@@ -1,1 +1,0 @@
-timestamp for config.h
